refactor ReleaseCard into components directory
