+++ services/frontend-service/src/assets/_components.scss
@@ -4,7 +4,7 @@
 @import 'src/ui/components/textfield/textfield';
 @import 'src/ui/components/navigation/navList';
 @import 'src/ui/NavigationBar/NavigationBar';
-@import 'src/ui/ReleaseCard/ReleaseCard';
+@import '../ui/components/ReleaseCard/ReleaseCard';
 @import 'src/ui/TopAppBar/TopAppBar';
 @import 'src/ui/components/navigation/navListItem';
 @import url('https://fonts.googleapis.com/icon?family=Material+Icons');
