+++ services/frontend-service/src/ui/components/ReleaseCard/ReleaseCard.scss
